+++ hydromt/model/processes/grid.py
@@ -42,8 +42,8 @@ def create_grid_from_region(
     crs: Optional[int] = None,
     region_crs: int = 4326,
     rotated: bool = False,
-    hydrography_fn: Optional[str] = None,
-    basin_index_fn: Optional[str] = None,
+    hydrography_path: Optional[str] = None,
+    basin_index_path: Optional[str] = None,
     add_mask: bool = True,
     align: bool = True,
     dec_origin: int = 0,
@@ -87,7 +87,7 @@ def create_grid_from_region(
         * Optional variables: ['basins'] if the `region` is based on a
             (sub)(inter)basins without a 'bounds' argument.
 
-    basin_index_fn : str, optional
+    basin_index_path: str, optional
         Name of data source with basin (bounding box) geometries associated with
         the 'basins' layer of `hydrography_fn`. Only required if the `region` is
         based on a (sub)(inter)basins without a 'bounds' argument.
@@ -158,12 +158,12 @@ def create_grid_from_region(
         geom = parse_region_basin(
             region,
             data_catalog=data_catalog,
-            hydrography_fn=hydrography_fn,
-            basin_index_fn=basin_index_fn,
+            hydrography_path=hydrography_path,
+            basin_index_path=basin_index_path,
         )
         xcoords, ycoords, crs = _extract_coords_from_basin(
             geom,
-            hydrography_fn=hydrography_fn,
+            hydrography_fn=hydrography_path,
             res=res,
             crs=crs,
             align=align,
@@ -279,7 +279,7 @@ def grid_from_constant(
     nodata: int, float, optional
         Nodata value. By default infered from dtype.
     mask_name: str, optional
-        Name of mask in self.grid to use for masking raster_fn. By default 'mask'.
+        Name of mask in self.grid to use for masking raster_data. By default 'mask'.
         Use None to disable masking.
 
     Returns
@@ -327,7 +327,7 @@ def grid_from_rasterdataset(
     ds: xr.DataArray, xr.Dataset
         Dataset with raster data.
     variables: list, optional
-        List of variables to add to grid from raster_fn. By default all.
+        List of variables to add to grid from raster_data. By default all.
     fill_method : str, optional
         If specified, fills nodata values using fill_nodata method.
         Available methods are {'linear', 'nearest', 'cubic', 'rio_idw'}.
@@ -335,11 +335,11 @@ def grid_from_rasterdataset(
         See rasterio.warp.reproject for existing methods, by default 'nearest'.
         Can provide a list corresponding to ``variables``.
     mask_name: str, optional
-        Name of mask in self.grid to use for masking raster_fn. By default 'mask'.
+        Name of mask in self.grid to use for masking raster_data. By default 'mask'.
         Use None to disable masking.
     rename: dict, optional
-        Dictionary to rename variable names in raster_fn before adding to grid
-        {'name_in_raster_fn': 'name_in_grid'}. By default empty.
+        Dictionary to rename variable names in raster_data before adding to grid
+        {'name_in_raster_data': 'name_in_grid'}. By default empty.
 
     Returns
     -------
@@ -396,17 +396,17 @@ def grid_from_raster_reclass(
     reclass_table: pd.DataFrame
         Tabular pandas dataframe object for the reclassification table of `da`.
     reclass_variables: list
-        List of reclass_variables from reclass_table_fn table to add to maps.
-        Index column should match values in `raster_fn`.
+        List of reclass_variables from reclass_table_data table to add to maps.
+        Index column should match values in `raster_data`.
     fill_method : str, optional
-        If specified, fills nodata values in `raster_fn` using fill_nodata method
+        If specified, fills nodata values in `raster_data` using fill_nodata method
         before reclassifying. Available methods are
         {'linear', 'nearest', 'cubic', 'rio_idw'}.
     reproject_method: str, optional
         See rasterio.warp.reproject for existing methods, by default "nearest".
         Can provide a list corresponding to ``reclass_variables``.
     mask_name: str, optional
-        Name of mask in self.grid to use for masking raster_fn. By default 'mask'.
+        Name of mask in self.grid to use for masking raster_data. By default 'mask'.
         Use None to disable masking.
     rename: dict, optional
         Dictionary to rename variable names in reclass_variables before adding to grid
@@ -475,19 +475,19 @@ def grid_from_geodataframe(
     gdf : gpd.GeoDataFrame
         geopandas object to rasterize.
     variables : List, str, optional
-        List of variables to add to grid from vector_fn. Required if
+        List of variables to add to grid from vector_data. Required if
         rasterize_method is "value", by default None.
     nodata : List, int, float, optional
         No data value to use for rasterization, by default -1. If a list is provided,
         it should have the same length has variables.
     rasterize_method : str, optional
         Method to rasterize the vector data. Either {"value", "fraction", "area"}.
-        If "value", the value from the variables columns in vector_fn are used
+        If "value", the value from the variables columns in vector_data are used
         directly in the raster. If "fraction", the fraction of the grid cell covered
         by the vector file is returned. If "area", the area of the grid cell covered
         by the vector file is returned.
     mask_name: str, optional
-        Name of mask in self.grid to use for masking raster_fn. By default 'mask'.
+        Name of mask in self.grid to use for masking raster_data. By default 'mask'.
         Use None to disable masking.
     rename: dict or str, optional
         Dictionary to rename variable names in variables before adding to grid
@@ -501,7 +501,7 @@ def grid_from_geodataframe(
     Returns
     -------
     ds_out: xr.Dataset
-        Dataset with data from vector_fn resampled to grid_like.
+        Dataset with data from vector_data resampled to grid_like.
     """  # noqa: E501
     # Check which method is used
     rename = rename or dict()